backend/requirements.txt
asgiref==3.8.1
async-timeout==5.0.1
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
cffi==1.17.1
channels==4.1.0
click==8.1.7
colorama==0.4.6
constantly==23.10.4
contourpy==1.3.1
cryptography==43.0.3
cycler==0.12.1
daphne==4.1.2
dj-database-url==2.3.0
Django==5.1.2
django-cors-headers==4.5.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
dlib==19.24.6
fonttools==4.55.5
hyperlink==21.0.0
idna==3.10
imutils==0.5.4
incremental==24.7.2
kiwisolver==1.4.8
lxml==5.3.0
matplotlib==3.10.0
msgpack==1.1.0
numpy==2.1.3
opencv-contrib-python==4.10.0.84
packaging==24.2
pillow==11.0.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.9.0
PyMuPDF==1.25.3
pyOpenSSL==24.2.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-magic-bin==0.4.14
pytz==2024.2
redis==5.2.0
scipy==1.14.1
service-identity==24.2.0
setuptools==75.3.0
six==1.17.0
sqlparse==0.5.1
tomli==2.0.2
Twisted==24.10.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.2
zope.interface==7.1.1
